Extract shared doc types and utils to core/docs
Moves the platform-neutral doc layer out of desktop/tabs/docs into a new core/docs module so the
mobile client can reuse it. Adds core/docs/types (doc entry/category/source/section types) and
core/docs/DocUtils (docRouteParams, resolveDocLink, normalizePath, extractSections, slugify). The
desktop docRegistry now holds only the desktop-route example map and re-exports the shared types and
resolveDocLink for back-compat; DocService and ToolboxLink import from the shared module. No behavior
change.
