+++ client-app/src/desktop/tabs/docs/docRegistry.ts
@@ -1,112 +1,17 @@
 /**
- * Types and utilities for the multi-source documentation viewer.
+ * Desktop-only doc -> Toolbox example tab mappings.
  *
- * The registry itself is loaded from the server (DocsService) at runtime.
- * This file retains the type definitions and utility functions used by
- * DocsPanelModel and DocsTab.
+ * Shared doc types and link/section utilities now live in `core/docs`; this file retains the
+ * desktop-route example map and re-exports the shared types + `resolveDocLink` so existing desktop
+ * importers keep resolving.
  */
-import {DocService} from '../../../core/svc/DocService';
+import {DocExampleLink} from '../../../core/docs/types';
+
+export type {DocEntry, DocCategory, DocSourceInfo, DocExampleLink} from '../../../core/docs/types';
+export {resolveDocLink} from '../../../core/docs/DocUtils';
 
 // ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-export interface DocEntry {
-    /** Unique identifier AND relative file path (e.g. 'docs/base-classes.md'). */
-    id: string;
-    source: string;
-    title: string;
-    category: string;
-    description: string;
-    keywords: string[];
-}
-
-export interface DocCategory {
-    id: string;
-    title: string;
-}
-
-export interface DocSourceInfo {
-    label: string;
-    categories: DocCategory[];
-    mode: string;
-}
-
-export interface DocExampleLink {
-    title: string;
-    /** Full Router5 route name, e.g. 'default.grids.standard'. */
-    route: string;
-}
-
-// ---------------------------------------------------------------------------
-// Link resolution
-// ---------------------------------------------------------------------------
-/**
- * Resolve a relative link from one doc to another.
- * Given the current doc entry and a relative href (e.g., '../core/README.md'),
- * returns the matching DocEntry, or undefined if not found.
- *
- * Since entry IDs are now file paths, we resolve the relative href against
- * the current doc's directory and look up the result directly in the registry.
- */
-export function resolveDocLink(currentDoc: DocEntry, href: string): DocEntry | undefined {
-    if (href.startsWith('http') || href.startsWith('#') || href.startsWith('mailto:')) {
-        return undefined;
-    }
-
-    const cleanHref = href.split('#')[0];
-    if (!cleanHref) return undefined;
-
-    const docService = DocService.instance;
-
-    // Resolve the relative path from the current doc's directory
-    const currentDir = currentDoc.id.substring(0, currentDoc.id.lastIndexOf('/') + 1);
-    const resolved = normalizePath(currentDir + cleanHref);
-
-    // Check within same source first
-    const sameSourceDoc = docService.registry.find(
-        e => e.source === currentDoc.source && e.id === resolved
-    );
-    if (sameSourceDoc) return sameSourceDoc;
-
-    // Check for cross-source links (e.g. ../../hoist-core/docs/authentication.md)
-    if (currentDoc.source === 'hoist-react') {
-        const coreMatch = resolved.match(/^(?:\.\.\/)*hoist-core\/(.+)$/);
-        if (coreMatch) {
-            return docService.registry.find(
-                e => e.source === 'hoist-core' && e.id === coreMatch[1]
-            );
-        }
-    }
-
-    if (currentDoc.source === 'hoist-core') {
-        const reactMatch = cleanHref.match(/^(?:\.\.\/)*hoist-react\/(.+)$/);
-        if (reactMatch) {
-            return docService.registry.find(
-                e => e.source === 'hoist-react' && e.id === reactMatch[1]
-            );
-        }
-    }
-
-    return undefined;
-}
-
-/** Normalize a path by resolving `.` and `..` segments. */
-function normalizePath(path: string): string {
-    const parts = path.split('/');
-    const result: string[] = [];
-    for (const part of parts) {
-        if (part === '.' || part === '') continue;
-        if (part === '..') {
-            result.pop();
-        } else {
-            result.push(part);
-        }
-    }
-    return result.join('/');
-}
-
-// ---------------------------------------------------------------------------
-// Doc → Toolbox example tab mappings (hoist-react only)
+// Doc -> Toolbox example tab mappings (hoist-react only)
 // ---------------------------------------------------------------------------
 const R = 'default';
